# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Lambda360 API",
    "version": "0.0.0"
  },
  "tags": [],
  "paths": {
    "/hello": {
      "get": {
        "operationId": "Hello_sayHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/step/{sha256}": {
      "get": {
        "operationId": "Step_exists",
        "parameters": [
          {
            "name": "sha256",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileExists"
                }
              }
            }
          }
        }
      }
    },
    "/view": {
      "get": {
        "operationId": "Viewer_view",
        "parameters": [
          {
            "name": "sha256",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "explode": false
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "model/gltf-binary": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "FileExists": {
        "type": "object",
        "required": [
          "exists"
        ],
        "properties": {
          "exists": {
            "type": "boolean"
          },
          "uploadUrl": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Main server",
      "variables": {}
    }
  ]
}